[FIX] stock: correctly compute custom search method
The search on qty_available was refactored twice:
- c64f5a12 to split the method
- eae63b43 to remove the eval operator

At c64f5a12 the order of evaluation introduced a bug so that the domain
   ('qty_available', '=' 10.0)
was evaluated
   if product[qty_available] = 10.0:
which crashed (`=` instead of `==`).

Before c64f5a12, this code did not crash as the operator was converted via
  if operator == '=':
    operator = '=='
which is no longer executed after the split

eae63b43 replaced the eval by a dictionnary that replaced the eval error by a
KeyError when searching with `=` operator.

This commit changes several things:
- replace the `operator in ('==', '>=', '<=')` as this was evaluated before the
  operator replacement code mentionned above
- remove the replacement code as no longer justified since eae63b43
- replace the OPERATOR key `==` to always match given key