[FIX] purchase: correct triggers on computed fields
To be triggered on the fields that are used inside the computed field.
Was not very problematic before as the fields were not stored and recomputed
at each modification. It could be backported if needed.

Next commit will store these fields for performance reason, fixing the trigger
first is required.

Linked to #15259