[FIX] web: do not call session load_qweb on frontend
Since commits 09a070ae720779893ac882499086aa95a62994d8 and c7646af04512dfbd0025c22c262e8edd25f5c194,
the session is not started by the webclient anymore but directly by the
session service instead, and the session info is found in the DOM
instead of by RPC.

Because of that, the session is started in the frontend and other
non-backend environments which caused a JS error while trying to
load qweb (as no qweb has to be loaded). This commit makes the
session start check if it is a non-backend environment and only
initialize session info (from DOM) in that case.