[FIX] account: register payment with payment, invoice and company having different currencies
In revision c681d6bf350b1f240f8a137756c77725f792b2d0,
`_compute_total_invoices_amount()` was assumed to
give the invoice residual within the company currency but,
in the case the payment_currency is different than the company_currency,
it returns the amount in the payment currency.

In this case, we actually need the total invoices residual
within the company currency.