[FIX] sale: Traceback product_packaging on SO line
When we add a SO line on a sale order with a product packaging, a
traceback is raised.

The problem comes from a default value and a domain set in the context
which reference a product template but since this rev: 64a76bde42502509add0d796ca6f1d56a83c38df
the product packagings are linked to product_product and not
product_template anymore.

So we used the product_id field instead of product_tmpl_id. And remove
product_tmpl_id from the sale_order_line model which is not used anymore