[FIX] account: company contact data
This is a complement of previous commit.

As a portal user, access one of your invoices from the frontend
(download the PDF). The generation of the report crashes because of an
access error. The error arises when trying to access the company banking
information displayed in the report footer.

opw-708581