[FIX] product_visible_discount: sale price for pricelist based on other pricelist
For a pricelist based on another pricelist,
the price of the product is the price in this other
pricelist.

To get the price of a product in a specific pricelist,
the field `price` must be used, along with the right `pricelist`
passed in the context when browsing the product.

Without this, the sale price is considered the sale price
indicated on the form, in the product company currency,
while the pricelist on which is based the currenct pricelist
could give a specific other price, in another specific currency.

e.g.:
Watch, Sale price 690CHF
Public USD Pricelist, setting the sale price of the watch to 790 USD
Reseller USD Pricelist, discounting 60% based on the public pricelist price, discount shown.

Before this revision, the sale price of the watch was marked
690 USD, with a sown discount of 54,xx %

With this revision, the sale price of the watch is marked at
790USD, with a shown discount of 60%, as expected.

opw-708301