[FIX] web: properly handles dirty data in form views
Before this commit, we only consider the BasicModel point of view to
determine if a record is dirty.  It is good when we want to determine if
it is actually necessary to save the data, but not correct when we want
to know if some changes can be discarded.

With this commit, the form controller keeps track of its 'dirty' state,
coming from user interactions.  It uses this flag to determine if
changes can be discarded.