[IMP] mail: remove duplicated logs when scheduling activities
Currently when scheduling activities, duplicate logs are created. Indeed
a message is posted when scheduling an activity, editing it and marking it
as done. This leads to a lot of noise being created on records.

This commit remove logs when scheduling and updating an activity. A sole
message is posted when the activity is marked as done.

This is a partial backport of commit e81e5af62075376e4afd45df24833a55b47dd380 targetting saas-15.