[FIX] sale_stock: don't link groupless backorders to all sales
When a backorder is created from an outgoing picking, a message is posted,
linking to the related sale orders as source of the original picking.

If the initial picking was not created from a sale, the resulting backorder
group was `False`.
The domain ('procurement_group_id', '=', False) matches many unrelated
(e.g. unconfirmed) sale orders.

Avoid trying to find the origin SO for pickings with no SO

Closes #15693